+++ src/restSpa/utils.nim
@@ -1,6 +1,15 @@
 # from std/logging import nil
-from times import getTime, toUnix
+from std/times import getTime, toUnix
+
 
 proc nowUnix*: int64 =
   ## Returns the unix time of now
   getTime().toUnix
+
+from std/uri import `$`
+from restSpa/config import address
+
+proc url*(path: string): string =
+  var link = address
+  link.path = path
+  result = $link